Only support HTML5 History API
This may be modified if there is a compelling reason to bring this back, but currently the only reason to not use this would be for stock Android browser for  Android <= v4.1. Opera Mini does not support it either, but it seems that Opera Mini breaks most (if not all?) JS, so it can be ignored.