Switch to babel in integration tests
buble wasn't working with Object.assign, so switch to babel since that is used everywhere else.

Also extended the setTimeout time in the unit tests from 10 to 100. This isn't typically a problem, but it seems like occasionally 10ms isn't a long enough timeout for the popstate event to be fired.