Don't test IE 11 for now
There is an issue that occurs with Internet Explorer and the hash history. I'm not sure exactly what the cause is, but it only seems to happen during Karma tests. Manually testing the same behavior, IE 11 works as expected. Will need to do some local digging instead of filling up Travis/BrowserStack with debug builds.