Rename "navigate" to "update"
This is something that a fair bit of time could be spent bikeshedding. I'm not sure what a really satisfying name for this would be, but I think that "update" is good enough. When you call history.update(nextLocation), you are updating the session history. "next" might work as well, but I feel that that sounds more like the functionality of "push" instead of either pushing or replacing.