Change argument signature passed to confirmation function
The first argument is now an object containing the location we are navigating to, the location we are navigating from, and the type of navigation. The second argument is the confirm function (previously called success, although the name didn't really matter) and the third argument is the prevent function (previously called failure).