Set initial action to PUSH (usually)
When creating a browser/hash history, we check if there is a history.state.key value set for the current location. If there is not, we set action to PUSH. If there is, we set it to POP.

The reason for this is that when a location POPs, we expect the browser to handle scroll restoration for us. However, when we are visiting a location for the first time, there is no scroll position to restore. This change will make it easier for us to know that scrolling should be handled manually (window.history.pushState does not auto-scroll, so we have to handle it ourselves for PUSH actions).