Add history.respondWith, remove subscriptions
Previously, a history instance would emit location changes to all of its subscribers. Now, it only has one "subscriber", which is referred to as a response handler. The response handler is responsible for any route loading as well as for actually updating the array of locations (native history in browsers, an array for in-memory)